Create editform.html
Note that this is a barebones version so I can write my JS for it - feel free to build on it if it fits your design. All QA fields are included, and I would prefer if IDs remain for inputs, textareas and buttons, as well as the classes for them, but I can change my JS to fit new ids if necessary. - Matt