Created barbones list
Barebones filters and table so I can start doing javascript without having to worry about filters or the table changing. I think this has all the fields we need for filters and columns for data we'll display when editing. I think it would be good to change colour to grey on hover for every row except the table head. Feel free to do whatever with the labels, I don't really need them. The create new can be a button or just text be in the header with it's own div, but I'll need to add an event listener to it so it needs a unique id